requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
python-dotenv==1.0.0
aiosqlite==0.20.0
asyncpg==0.29.0
apscheduler==3.10.4
